Support links opening a new window
Fixes DEV-16193
